requirements.txt
asgiref==3.5.2
atomicwrites==1.4.1
attrs==21.4.0
backports.zoneinfo==0.2.1
colorama==0.4.5
coverage==6.4.2
Django==4.0.6
django-filter==22.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
iniconfig==1.1.1
model-bakery==1.6.0
packaging==21.3
pluggy==1.0.0
psycopg2==2.9.3
py==1.11.0
PyJWT==2.4.0
pyparsing==3.0.9
pytest==7.1.2
pytest-cov==3.0.0
pytest-django==4.5.2
python-dotenv==0.20.0
pytz==2022.1
sqlparse==0.4.2
tomli==2.0.1
tzdata==2022.1
